pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "autorl-loop"
version = "0.1.0"
description = "Universal AutoRL hyperparameter search framework with LLM-driven decision making"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
keywords = ["reinforcement-learning", "hyperparameter-search", "automl", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.20.0"]
all = ["anthropic>=0.20.0"]

[project.scripts]
autorl = "autorl.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/autorl-loop"
Issues = "https://github.com/yourusername/autorl-loop/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["autorl*"]
